refactor(today): extract presentation helpers
